mildoc_index/requirements.txt
# Minio对象存储客户端
minio==7.2.20

# Milvus向量数据库客户端
pymilvus==2.6.5

# OpenAI API客户端（用于embedding）
openai==2.11.0

# 环境变量管理
python-dotenv==1.2.1

# PDF文档解析
PyPDF2==3.0.1

# Office文档解析
markitdown[all]==0.1.4

# text splitters
langchain-text-splitters==1.1.0
# langchain-core==1.2.0


# 阿里云
alibabacloud-oss-v2==1.2.1
aliyun-mns-sdk==1.3.0